Removing unneeded mapping
